add geometry submodule to raytracing
