+++ artist/__init__.py
@@ -2,7 +2,7 @@
 
 import os
 
-from . import core, field, io, scenario, scene, util
+from . import core, field, io, nurbs, raytracing, scenario, scene, util
 
 ARTIST_ROOT = f"{os.sep}".join(__file__.split(os.sep)[:-2])
 """Reference to the root directory of ``ARTIST``."""
@@ -11,6 +11,8 @@ __all__ = [
     "core",
     "io",
     "field",
+    "nurbs",
+    "raytracing",
     "scenario",
     "scene",
     "util",
